+++ R/INARtest.R
@@ -1,6 +1,6 @@
-#' Perform Sun-McCabe INAR tests
+#' Perform Sun-McCabe INAR(1) tests.
 #'
-#' @param X vector, thinning parameters. The length of this vector defines the number of lags `p` of the INAR(p) process.
+#' @param X vector, the observed series of counts.
 #' @param inn character, the innovation distribution to be used in the test. The options are "poi", "negbin" and "genpoi".
 #' @param B integer, the number of bootstrap samples to be generated. The default value is 0, corresponding to no bootstrap samples.
 #' @param method character, the type of test to be performed, the alternatives are "par" or "semipar".
@@ -43,6 +43,8 @@ SMCtest <- function(X, inn = "poi", B = 0, method = NA, saveboot = FALSE){
     OUT$data.name = deparse1(substitute(X)) # deparse(args(X))
 
     smc_est <- SMC_Cpp(X, OUT$inn_num)
+    OUT$statistic <- c(S = smc_est[1])
+    OUT$p.value <- smc_est[2]
 
     if(B > 0){
         # perform bootstrap test
@@ -55,15 +57,11 @@ SMCtest <- function(X, inn = "poi", B = 0, method = NA, saveboot = FALSE){
         }
 
         # add Boot results to OUT
-        OUT$statistic <- c(Sb = mean(smc_boot))
-        OUT$p.value <- mean(abs(smc_boot) > abs(smc_est[1]), na.rm = TRUE)
-        OUT$statistic.exact <- c(S = smc_est[1])
-        OUT$p.value.exact <- smc_est[2]
+        OUT$statistic.boot <- c(Sb = mean(smc_boot))
+        OUT$p.value.boot <- mean(abs(smc_boot) > abs(smc_est[1]), na.rm = TRUE)
         if(saveboot) OUT$bootvec <- smc_boot
-    }else{
-        OUT$statistic <- c(S = smc_est[1])
-        OUT$p.value <- smc_est[2]
     }
+
     return(OUT)
 }
 
@@ -113,8 +111,8 @@ DItest <- function(X){ # , inn = "poi"
 }
 
 # DItest(rpois(1000,2))
-# DItest(genINAR(1000,a = 0.5, par = 2,arrival = "poisson")$X)
-# DItest(genINAR(1000,a = 0.5, par = c(2,0.7),arrival = "negbin")$X)
+# DItest(genINAR(1000,a = 0.5, par = 2,inn = "poisson")$X)
+# DItest(genINAR(1000,a = 0.5, par = c(2,0.7),inn = "negbin")$X)
 
 
 #' Perform the zero inflation test for the INAR(1) (Weiss et al, 2019)
@@ -172,10 +170,10 @@ ZItest <- function(X,type = "pv"){ # , inn = "poi"
 # ZItest(rpois(1000,2),"pv")
 # ZItest(rpois(1000,2),"vdb")
 #
-# ZItest(genINAR(1000,a = 0.5, par = 2,arrival = "poisson")$X,"pv")
-# ZItest(genINAR(1000,a = 0.5, par = 2,arrival = "poisson")$X,"vdb")
-# ZItest(genINAR(1000,a = 0.5, par = c(2,0.7),arrival = "negbin")$X,"pv")
-# ZItest(genINAR(1000,a = 0.5, par = c(2,0.7),arrival = "negbin")$X,"vdb")
+# ZItest(genINAR(1000,a = 0.5, par = 2,inn = "poisson")$X,"pv")
+# ZItest(genINAR(1000,a = 0.5, par = 2,inn = "poisson")$X,"vdb")
+# ZItest(genINAR(1000,a = 0.5, par = c(2,0.7),inn = "negbin")$X,"pv")
+# ZItest(genINAR(1000,a = 0.5, par = c(2,0.7),inn = "negbin")$X,"vdb")
 
 
 #
@@ -242,77 +240,61 @@ ZIDItest <- function(X, type = "pv"){
 # ZIDItest(rpois(1000,2),"pv")
 # ZIDItest(rpois(1000,2),"vdb")
 #
-# ZIDItest(genINAR(1000,a = 0.5, par = 2,arrival = "poisson")$X,"pv")
-# ZIDItest(genINAR(1000,a = 0.5, par = 2,arrival = "poisson")$X,"vdb")
+# ZIDItest(genINAR(1000,a = 0.5, par = 2,inn = "poisson")$X,"pv")
+# ZIDItest(genINAR(1000,a = 0.5, par = 2,inn = "poisson")$X,"vdb")
 # ZIDItest(genINAR(1000,a = 0.5, par = c(2,0.9),inn = "negbin")$X,"pv")
 # ZIDItest(genINAR(1000,a = 0.5, par = c(2,0.1),inn = "negbin")$X,"vdb")
 
-# #' Perform Harrison-McCabe INAR tests
-# #'
-# #' @param X vector, thinning parameters. The length of this vector defines the number of
-# #' lags `p` of the INAR(p) process.
-# #' @param inn character, the innovation distribution to be used in the test.
-# #' The options are "poi", "negbin" and "genpoi".
-# #' @param B integer, the number of bootstrap samples to be generated.
-# #' The default value is 0, corresponding to no bootstrap samples.
-# #' @param method character, the type of test to be performed,
-# #' the alternatives are "par" or "semipar".
-# #' @return A number.
-# #'
-# #' @details
-# #' The function performs the Harrison-McCabe (HMC) tests for the INAR
-# #' model. It is possible to run an exact (B = 0) or a bootstrap (B > 0) test.
-# #' HMC tests are based on Poisson, negative binomial and generalized Poisson
-# #' distributions, in all the considered cases, both parametric and semiparametric
-# #' methods are available.
-# #' It is possible to perform the exact tests using the original data (B=0)
-# #' or using bootstrap samples (B > 0).
-# #' @examples
-# #' # ....... examples .....
-# #' # HMCtest(X = rpois(100,2), type = "semiparametric", B = 0)
-# #'
-# #' @export
-# #'
-# HMCtest <- function(X, inn = "poi", B = 0, method = NA){
-#     if(!all(X == as.integer(X))) X <- as.integer(X)
-#     if(!is.integer(B)) B <- as.integer(B)
-#
-#     stopifnot(inn %in% info_inn$inn, B >= 0)
-#     stopifnot(is.na(method) | method %in% c("par","semipar"))
-#     if(B > 0 & is.na(method)) stop("Specify the method: 'par' or 'semipar'")
-#     stopifnot(all(X == as.integer(X)), !all(X == X[1]), is.integer(B))
-#     if(inn == "negbin" & var(X) <= mean(X)){
-#         stop("Underdispersion detected: Negative Binomial test not applicable (var <= mean).\n Try switching to generalized Poisson distribution.");
-#     }
-#
-#     OUT <- get_info(list(
-#         test = "hmc",
-#         inn = inn,
-#         method = method,
-#         B = B
-#     ))
-#
-#     if(B > 0 & method == "par" & !OUT$check) stop(paste("The selected innovation distribution (",inn,") is not available for the parametric HMC bootstrap test",sep=""))
-#     OUT$data.name = deparse1(substitute(X)) # deparse(args(X))
-#
-#     hmc_est <- HMC_Cpp(X, OUT$inn_num)
-#
-#     if(B > 0){
-#         # perform bootstrap test
-#         if(method == "par"){
-#             hmc_boot <- HMC_parBOOT_Cpp(X, B, OUT$inn_num)
-#         }else if(method == "semipar"){
-#             hmc_boot <- HMC_semiparBOOT_Cpp(X, B, OUT$inn_num)
-#         }else{
-#             stop("Specify a correct test type: 'par' or 'semipar'")
-#         }
-#
-#         # add Boot results to OUT
-#         OUT$statistic <- c(T = mean(hmc_boot))
-#         OUT$p.value <- mean(abs(hmc_boot) > abs(hmc_est[1]), na.rm = TRUE)
-#     }else{
-#         OUT$statistic <- c(T = hmc_est[1])
-#         OUT$p.value <- hmc_est[2]
-#     }
-#     return(OUT)
-# }
+
+#' Perform Harris-McCabe INAR(1) test.
+#'
+#' @param X vector, the observed series of counts.
+#' @param B integer, the number of bootstrap samples to be generated.
+#' @param saveboot logical, if TRUE the bootstrap replicates are saved in the output object (only if B > 0).
+#' The default value is 0, corresponding to no bootstrap samples.
+#' @return A number.
+#'
+#' @details
+#' The function performs the Harris-McCabe (HMC) tests for the INAR
+#' model. It is possible to run an exact (B = 0) or a bootstrap (B > 0) test.
+#' It is possible to perform the exact test using the original data (B=0)
+#' or using bootstrap samples (B > 0).
+#' @examples
+#' # ....... examples .....
+#' # HMCtest(X = rpois(100,2), B = 0)
+#'
+#' @export
+HMCtest <- function(X, B = 0, saveboot = FALSE){
+    if(!all(X == as.integer(X))) X <- as.integer(X)
+    if(!is.integer(B)) B <- as.integer(B)
+
+    stopifnot(all(X == as.integer(X)), !all(X == X[1]), is.integer(B))
+
+    OUT <- get_info(list(
+        test = "hmc",
+        inn = "none",
+        method = NA,
+        B = B
+    ))
+
+    OUT$data.name = deparse1(substitute(X)) # deparse(args(X))
+
+    hmc_est <- HMC_Cpp(X)
+    OUT$statistic <- c(S = hmc_est[1])
+    OUT$p.value <- hmc_est[2]
+
+    if(B > 0){
+        hmc_boot <- HMC_BOOT_Cpp(X, B)
+
+        # add Boot results to OUT
+        OUT$statistic.boot <- c(Sb = mean(hmc_boot))
+        OUT$p.value.boot <- mean(abs(hmc_boot) > abs(hmc_est[1]), na.rm = TRUE)
+        if(saveboot) OUT$bootvec <- hmc_boot
+    }
+
+    return(OUT)
+}
+
+
+# HMCtest(rpois(1000,2), B = 0)
+# HMCtest(rpois(1000,2), B = 399)